Add GradCam visualization
Add submodule pytorch-cnn-visualizations

Add code/visualizer/GradCamVisualizer.py. Starting this script with the path to a file
with the pytorch weights of the model as a command-line parameter will generate a GradCam
visualization for all convolutional layers of the model.
Caveat: Currently only implemented for CNN_5 (hardcoded layer numbers). Should parse this from
the model in the long term.